Addeb bypass for Android versioner
