+++ Extensions/Versioning/VersionAndroidManifest/src/ApplyVersionToManifest.ts
@@ -1,7 +1,8 @@
 import {
     getSplitVersionParts,
     updateManifestFile,
-    findFiles
+    findFiles,
+    extractVersion
 } from "./ApplyVersionToManifestFunctions";
 
 import tl = require("vsts-task-lib/task");
@@ -14,6 +15,7 @@ var versionNameFormat = tl.getInput("VersionNameFormat");
 var versionCodeFormat = tl.getInput("VersionCodeFormat");
 var outputversion = tl.getInput("outputversion");
 var filenamePattern = tl.getInput("FilenamePattern");
+var injectversion = tl.getBoolInput("Injectversion");
 
 console.log (`Source Directory:  ${path}`);
 console.log (`Filename Pattern: ${filenamePattern}`);
@@ -21,6 +23,7 @@ console.log (`Version Number/Build Number: ${versionNumber}`);
 console.log (`Version Filter to extract build number: ${versionRegex}`);
 console.log (`Version Number Format: ${versionNameFormat}`);
 console.log (`Version Code Format: ${versionCodeFormat}`);
+console.log (`Inject Version: ${injectversion}`);
 console.log (`Output: Version Number Parameter Name: ${outputversion}`);
 
 // Make sure path to source code directory is available
@@ -33,27 +36,7 @@ if (!fs.existsSync(path)) {
 var files = findFiles(path, filenamePattern, files);
 console.log (`Found ${files.length} files to update.`);
 
-var regexp = new RegExp(versionRegex);
-var versionData = regexp.exec(versionNumber);
-if (!versionData) {
-    // extra check as we don't get zero size array but a null
-    tl.error(`Could not find version number data in ${versionNumber} that matches ${versionRegex}.`);
-    process.exit(1);
-}
-switch (versionData.length) {
-   case 0:
-         // this is trapped by the null check above
-         tl.error(`Could not find version number data in ${versionNumber} that matches ${versionRegex}.`);
-         process.exit(1);
-   case 1:
-        break;
-   default:
-         tl.warning(`Found more than instance of version data in ${versionNumber}  that matches ${versionRegex}.`);
-         tl.warning(`Will assume first instance is version.`);
-         break;
-}
-
-const newVersion = versionData[0];
+const newVersion = extractVersion(injectversion, versionRegex, versionNumber);
 console.log (`Extracted Version: ${newVersion}`);
 
 const versionName = getSplitVersionParts(versionRegex, versionNameFormat, newVersion);
