+++ Extensions/Versioning/VersionAndroidManifest/test/android.tests.ts
@@ -5,7 +5,8 @@ import "mocha";
 import {
     getSplitVersionParts,
     updateManifestFile,
-    findFiles
+    findFiles,
+    extractVersion
 } from "../src/ApplyVersionToManifestFunctions";
 
 import fs = require("fs");
@@ -81,6 +82,20 @@ describe("Test the file processing", () => {
     after(function() {
       // remove the file if created
       del.sync("test/testdata/sample.xml");
+    });
+
+  });
+
+  describe("Test the version extraction", () => {
+
+    it("should be able to extract just a version for a build number", () => {
+      var actual = extractVersion(false, "\\d+\\.\\d+\\.\\d+", "ABC-1.2.3.4-XYZ");
+      expect(actual).equals("1.2.3");
+    });
+
+    it("should be able to skip extracting a version for a build number", () => {
+      var actual = extractVersion(true, "\\d+\\.\\d+\\.\\d+", "ABC-1.2.3.4-XYZ");
+      expect(actual).equals("ABC-1.2.3.4-XYZ");
     });
 
   });
